# server/.api/apis/eupago/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "REST API - Body Auth",
    "version": "unknown"
  },
  "servers": [
    {
      "url": "https://sandbox.eupago.pt/clientes/rest_api"
    }
  ],
  "components": {
    "securitySchemes": {}
  },
  "security": [
    {}
  ],
  "paths": {
    "/multibanco/create": {
      "post": {
        "summary": "Multibanco",
        "description": "Creates Multibanco payment references:\n- with or without payment deadline\n- define an amount or an interval with max/min amounts to receive the payment\n- may choose if the references allow only 1 or more that 1 payment",
        "operationId": "multibanco",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "chave",
                  "valor",
                  "per_dup"
                ],
                "properties": {
                  "chave": {
                    "type": "string",
                    "description": "API Key"
                  },
                  "valor": {
                    "type": "number",
                    "description": "Amount",
                    "format": "float"
                  },
                  "id": {
                    "type": "string",
                    "description": "Identifier"
                  },
                  "data_inicio": {
                    "type": "string",
                    "description": "Reference starts to allow payments on this day (YYYY-MM-DD)",
                    "format": "date"
                  },
                  "data_fim": {
                    "type": "string",
                    "description": "Payment deadline (YYYY-MM-DD)",
                    "format": "date"
                  },
                  "valor_maximo": {
                    "type": "number",
                    "description": "Maximum amount (only for references that allow payments in an interval of amounts)",
                    "format": "float"
                  },
                  "valor_minimo": {
                    "type": "number",
                    "description": "Minimum amount (only for references that allow payments in an interval of amounts)",
                    "format": "float"
                  },
                  "per_dup": {
                    "type": "integer",
                    "description": "Defines if the reference allows 1 payment or more that 1 payment (1 = allows multiple payments | 0 = allows only 1 payment)",
                    "format": "int32"
                  },
                  "campos_extra": {
                    "type": "array",
                    "description": "Use this object to assign extra fields. Note that extra fields must be first setted up on your account channel.",
                    "items": {
                      "properties": {
                        "id": {
                          "type": "integer",
                          "description": "Number of your extra field",
                          "format": "int32"
                        },
                        "valor": {
                          "type": "string",
                          "description": "Value that should be assigned to this extra field for this reference."
                        }
                      },
                      "required": [
                        "id",
                        "valor"
                      ],
                      "type": "object"
                    }
                  },
                  "failOver": {
                    "type": "string",
                    "description": "Send a reminder to your customer that their payment has expired (Only available on references with deadline).",
                    "enum": [
                      "0",
                      "1"
                    ]
                  },
                  "email": {
                    "type": "string",
                    "description": "User email to receive the reminder notification (Only if failover = 1)."
                  },
                  "contacto": {
                    "type": "string",
                    "description": "User phone number to receive the reminder notification (Only if failover = 1)."
                  },
                  "userID": {
                    "type": "string",
                    "description": "User identifier parameter. (Mandatory for some clients)"
                  }
                }
              },
              "examples": {
                "Without Deadline - Interval of Amounts": {
                  "value": {
                    "chave": "xxxx-xxxx-xxxx-xxxx-xxxx",
                    "valor": "0",
                    "id": "Exemplo-em-JSON",
                    "valor_minimo": "5",
                    "valor_maximo": "122.5",
                    "per_dup": "0"
                  }
                },
                "With Deadline - Specific Amount": {
                  "value": {
                    "chave": "xxxx-xxxx-xxxx-xxxx-xxxx",
                    "valor": "10.5",
                    "id": "Example-em-JSON",
                    "data_inicio": "2021-10-28",
                    "data_fim": "2021-11-01",
                    "per_dup": "0"
                  }
                },
                "Extra Fields": {
                  "value": {
                    "campos_extra": [
                      {
                        "id": 369,
                        "valor": "um valor"
                      },
                      {
                        "id": 178,
                        "valor": "outro valor"
                      }
                    ],
                    "chave": "xxxx-xxxx-xxxx-xxxx-xxxx",
                    "valor": "3",
                    "id": "Exemplo2",
                    "per_dup": "0"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "200",
            "content": {
              "application/json": {
                "examples": {
                  "Success": {
                    "value": "{\n    \"sucesso\": true,\n    \"estado\": 0,\n    \"resposta\": \"OK\",\n    \"referencia\": \"123456789\",\n    \"valor\": \"0.00000\",\n    \"entidade\": \"12345\",\n    \"valor_minimo\": \"5\",\n    \"valor_maximo\": \"122.5\",\n    \"data_inicio\": \"2021-10-28\",\n    \"data_fim\": \"2099-12-31\"\n}"
                  },
                  "Invalid API Key": {
                    "value": "{\n    \"sucesso\": false,\n    \"estado\": -10,\n    \"resposta\": \"Chave de API inválida.\"\n}"
                  },
                  "Invalid Data": {
                    "value": "{\n    \"sucesso\": false,\n    \"estado\": -10,\n    \"resposta\": \"Os dados inseridos não são válidos\"\n}"
                  }
                },
                "schema": {
                  "oneOf": [
                    {
                      "title": "Success",
                      "type": "object",
                      "properties": {
                        "sucesso": {
                          "type": "boolean",
                          "example": true,
                          "default": true
                        },
                        "estado": {
                          "type": "integer",
                          "example": 0,
                          "default": 0
                        },
                        "resposta": {
                          "type": "string",
                          "example": "OK"
                        },
                        "referencia": {
                          "type": "string",
                          "example": "123456789"
                        },
                        "valor": {
                          "type": "string",
                          "example": "0.00000"
                        },
                        "entidade": {
                          "type": "string",
                          "example": "12345"
                        },
                        "valor_minimo": {
                          "type": "string",
                          "example": "5"
                        },
                        "valor_maximo": {
                          "type": "string",
                          "example": "122.5"
                        },
                        "data_inicio": {
                          "type": "string",
                          "example": "2021-10-28"
                        },
                        "data_fim": {
                          "type": "string",
                          "example": "2099-12-31"
                        }
                      }
                    },
                    {
                      "title": "Invalid API Key",
                      "type": "object",
                      "properties": {
                        "sucesso": {
                          "type": "boolean",
                          "example": false,
                          "default": true
                        },
                        "estado": {
                          "type": "integer",
                          "example": -10,
                          "default": 0
                        },
                        "resposta": {
                          "type": "string",
                          "example": "Chave de API inválida."
                        }
                      }
                    },
                    {
                      "title": "Invalid Data",
                      "type": "object",
                      "properties": {
                        "sucesso": {
                          "type": "boolean",
                          "example": false,
                          "default": true
                        },
                        "estado": {
                          "type": "integer",
                          "example": -10,
                          "default": 0
                        },
                        "resposta": {
                          "type": "string",
                          "example": "Os dados inseridos não são válidos"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "deprecated": false,
        "security": []
      }
    },
...
  },
...
}